# [EX-1181] use uint256 for curve-base