… go.sum
… pkg/source/curve/base/math.go
… pkg/source/curve/base/pool_simulator_compare_test.go