Benchmark:

Before:
`BenchmarkCalcAmountOut-8   	  167113	      6275 ns/op	    5380 B/op	     109 allocs/op`

`BenchmarkGetDyUnderlying-8   	  136398	      8083 ns/op	    6645 B/op	     136 allocs/op`

After:
`BenchmarkCalcAmountOut-8   	  795174	      1507 ns/op	     240 B/op	       7 allocs/op`

`BenchmarkGetDyUnderlying-8   	  520291	      2309 ns/op	     816 B/op	      23 allocs/op`

## Why did we need it?
<!--- Describe your changes in detail -->

## Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

## Release Note
<!--- Please list out special notes (if any) when releasing this commit -->
<!--- Special notes is things like breaking changes and related components, required changes and env configurations in kyber-application, etc. -->

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):